% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{charlson_list}
\alias{charlson_list}
\title{List of Charlson comorbidities}
\format{A list, with one column for each comorbidity; value is a textual description}
\usage{
charlson_list
}
\description{
List of Charlson comorbidities
}
\examples{
# List the strings used to refer to Charlson comorbidities
names(charlson_list)

# List descriptions of comorbidities
charlson_list
}
\keyword{datasets}
