% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elixhauser_list}
\alias{elixhauser_list}
\title{List of Elixhauser comorbidities}
\format{A list, with one column for each comorbidity; value is a textual description}
\usage{
elixhauser_list
}
\description{
List of Elixhauser comorbidities
}
\examples{
# List the strings used to refer to Elixhauser comorbidities
names(elixhauser_list)

# List descriptions of comorbidities
elixhauser_list
}
\keyword{datasets}
