% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rsi_beta_inhosp}
\alias{rsi_beta_inhosp}
\title{Values for calculating RSI for in-hospital mortality.}
\format{A hash (see package "hash"), where key is icd9cm code, and value is beta.
Special key "popbeta" has the population beta for the entire table.}
\usage{
rsi_beta_inhosp
}
\description{
Values for calculating RSI for in-hospital mortality.
}
\references{
1. Sessler DI, Sigl JC, Manberg PJ, Kelley SD, Schubert A, Chamoun NG.
    Broadly applicable risk stratification system for predicting duration of hospitalization and mortality.
    Anesthesiology. 2010 Nov;113(5):1026-37. doi: 10.1097/ALN.0b013e3181f79a8d.
\url{http://my.clevelandclinic.org/anesthesiology/outcomes-research/risk-stratification-index.aspx}
}
