% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rsi_sample_data}
\alias{rsi_sample_data}
\title{Sample data for validating RSI}
\format{A data table with a patient ID and several columns with ICD-9-CM codes.}
\usage{
rsi_sample_data
}
\description{
Sample data for validating RSI
}
\references{
1. Sessler DI, Sigl JC, Manberg PJ, Kelley SD, Schubert A, Chamoun NG.
    Broadly applicable risk stratification system for predicting duration of hospitalization and mortality.
    Anesthesiology. 2010 Nov;113(5):1026-37. doi: 10.1097/ALN.0b013e3181f79a8d.
\url{http://my.clevelandclinic.org/anesthesiology/outcomes-research/risk-stratification-index.aspx}
}
\seealso{
\code{\link{rsi_sample_results}}, \code{\link{verify_sessler_rsi}}
}
