% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-rsi.R
\name{verify_sessler_rsi}
\alias{verify_sessler_rsi}
\title{Validates this Sessler RSI implementation against reference data}
\usage{
verify_sessler_rsi()
}
\value{
Table of patients with scores >0.001 difference.
}
\description{
Requires that "sample data rev2.csv" and "sample results rev2.csv" be available
in datasrc directory
}
\examples{
\dontrun{
verify_sessler_rsi()
}
}
\references{
1. Sessler DI, Sigl JC, Manberg PJ, Kelley SD, Schubert A, Chamoun NG.
    Broadly applicable risk stratification system for predicting duration of hospitalization and mortality.
    Anesthesiology. 2010 Nov;113(5):1026-37. doi: 10.1097/ALN.0b013e3181f79a8d.
}
\seealso{
rsi_sample_data, rsi_sample_results
}
\author{
Patrick McCormick <patrick.mccormick@mountsinai.org>
}
