% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vt_inp_sample}
\alias{vt_inp_sample}
\title{First 100 patients and their ICD-9-CM codes from the Vermont Uniform Hospital
Discharge Data Set for 2011, Inpatient.}
\format{A data frame, with column "id" (numeric), some descriptive columns,
  "dx" (factor), and "icd9cm" (factor)}
\usage{
vt_inp_sample
}
\description{
Diagnostic ICD-9 codes are prefixed with 'D', while procedural ICD-9 codes are prefixed with 'P'.
}
\examples{
max(vt_inp_sample$scu_days)
}
\references{
\url{http://healthvermont.gov/research/hospital-utilization/RECENT_PU_FILES.aspx}
}
\seealso{
\code{\link{icd9cm_charlson_deyo}}, 
\code{\link{icd9cm_charlson_quan}}, 
\code{\link{icd9cm_charlson_romano}}, 
\code{\link{icd9cm_elixhauser_quan}}, 
\code{\link{icd9cm_elixhauser_ahrq37}}
}
\keyword{datasets}
